Supported .ocd versions, hint on 'Cutout' usage
Update of the .ocd versions that can be read (2018 basically works, although issuing a warning) and can be written (8..12).
A hint was added when cutting parts of a map away that the original map can easily be restored by undoing. This might be important to a) reduce the file size and b) if someone gives a map away but does not want to give the whole map away.